requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
_tflow_select=2.3.0=eigen
absl-py=0.12.0=pyhd8ed1ab_0
aiohttp=3.7.4=py38h294d835_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blinker=1.4=py_1
boost-cpp=1.74.0=h54f0996_2
branca=0.4.2=pyhd8ed1ab_0
brotlipy=0.7.0=py38h294d835_1001
bzip2=1.0.8=h8ffe710_4
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.2.1=pyhd8ed1ab_0
cairo=1.16.0=hb19e0ff_1008
certifi=2020.12.5=py38haa244fe_1
cffi=1.14.5=py38hd8c33c5_0
cfitsio=3.470=h0af3d06_7
chardet=4.0.0=py38haa244fe_1
click=7.1.2=pyh9f0ad1d_0
click-plugins=1.1.1=py_0
cligj=0.7.1=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py38hd7da0ea_0
curl=7.76.1=hf1763fc_1
cycler=0.10.0=py_2
decorator=5.0.7=pyhd8ed1ab_0
descartes=1.1.0=py_4
expat=2.3.0=h39d44d4_0
fiona=1.8.19=py38h4bd599a_0
folium=0.12.0=pyhd8ed1ab_1
fontconfig=2.13.1=h1989441_1005
freetype=2.10.4=h546665d_1
freexl=1.0.6=ha8e266a_0
gast=0.4.0=pyh9f0ad1d_0
gdal=3.2.2=py38hacca965_2
geopandas=0.9.0=pyhd8ed1ab_0
geos=3.9.1=h39d44d4_2
geotiff=1.6.0=h8e90983_5
gettext=0.19.8.1=h1a89ca6_1005
google-auth=1.28.0=pyh44b312d_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.37.0=py38he5377a8_0
h5py=2.10.0=nompi_py38he6c2248_106
hdf4=4.2.13=h0e5069d_1005
hdf5=1.10.6=nompi_h5268f04_1114
icu=68.1=h0e60522_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.10.1=py38haa244fe_0
intel-openmp=2021.2.0=h57928b3_616
ipykernel=5.5.3=py38h43734a8_0
ipython=7.22.0=py38h43734a8_0
ipython_genutils=0.2.0=py_1
jedi=0.18.0=py38haa244fe_2
jinja2=2.11.3=pyh44b312d_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h8ffe710_0
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38haa244fe_0
kealib=1.4.14=h96bfa42_2
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py38hbd9d945_1
krb5=1.17.2=hbae68bd_0
lcms2=2.12=h2a16943_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libcurl=7.76.1=hf1763fc_1
libffi=3.3=h0e60522_2
libgdal=3.2.2=h48ef39c_2
libglib=2.68.1=h1e62bf3_0
libiconv=1.16=he774522_0
libkml=1.3.0=h9859afa_1013
liblapack=3.9.0=8_mkl
libnetcdf=4.7.4=nompi_h3a9aa94_107
libpng=1.6.37=h1d00b33_2
libpq=13.1=h4f54205_2
libprotobuf=3.15.8=h7755175_0
librttopo=1.1.0=hb340de5_6
libsodium=1.0.18=h8d14728_1
libspatialindex=1.9.3=h39d44d4_3
libspatialite=5.0.1=h6b539a6_4
libssh2=1.9.0=h680486a_6
libtiff=4.2.0=hc10be44_0
libwebp-base=1.2.0=h8ffe710_2
libxml2=2.9.10=hf5bbc77_4
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=py38h294d835_3
matplotlib-base=3.4.1=py38heae8d8c_0
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
multidict=5.1.0=py38h294d835_1
munch=2.5.0=py_0
networkx=2.3=py_0
numpy=1.20.2=py38h09042cb_0
oauthlib=3.0.1=py_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=h48faf41_0
openssl=1.1.1k=h8ffe710_0
opt_einsum=3.3.0=pyhd8ed1ab_1
osmnx=1.0.1=pyhd3deb0d_0
pandas=1.2.4=py38h60cbd38_0
parso=0.8.2=pyhd8ed1ab_0
pcre=8.44=ha925a31_0
pickleshare=0.7.5=py_1003
pillow=8.1.2=py38h9273828_1
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h8ffe710_0
poppler=21.03.0=h9ff6ed8_0
poppler-data=0.4.10=0
postgresql=13.1=h0f1a9bc_2
proj=8.0.0=h1cfcee9_0
prompt-toolkit=3.0.18=pyha770c72_0
protobuf=3.15.8=py38h885f38d_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.8.1=pyhd8ed1ab_0
pyjwt=2.0.1=pyhd8ed1ab_1
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=3.0.1=py38he394cdd_1
pyreadline=2.1=py38haa244fe_1003
pysocks=1.7.1=py38haa244fe_3
python=3.8.8=h7840368_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2021.1=pyhd8ed1ab_0
pywin32=300=py38h294d835_0
pyzmq=22.0.3=py38h09162b1_1
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7.2=pyh44b312d_0
rtree=0.9.7=py38h8b54edf_1
scikit-learn=0.24.1=py38ha09990b_0
scipy=1.6.2=py38he847743_0
setuptools=49.6.0=py38haa244fe_3
shapely=1.7.1=py38h2426642_4
six=1.15.0=pyh9f0ad1d_0
sqlite=3.35.4=h8ffe710_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow=2.3.0=mkl_py38h8c0d9a2_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.4.0=pyh9656e83_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tiledb=2.2.7=hf84e3da_0
tk=8.6.10=h8ffe710_1
tornado=6.1=py38h294d835_1
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pyhd8ed1ab_0
vc=14.2=hb210afc_4
vs2015_runtime=14.28.29325=h5e1d092_4
wcwidth=0.2.5=pyh9f0ad1d_2
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38haa244fe_1006
wrapt=1.12.1=py38h294d835_3
xerces-c=3.2.3=h0e60522_2
xz=5.2.5=h62dcd97_1
yarl=1.6.3=py38h294d835_1
zeromq=4.3.4=h0e60522_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
